fixed mentor remarks
