+++ model/contacts/listContacts.js
@@ -1,0 +1,5 @@
+const contacts = require("../../db/contacts.json");
+
+const listContacts = async () => contacts;
+
+module.exports = listContacts;
