+++ services/randomId.js
@@ -1,7 +1,7 @@
-function getRandomInt(min = 11, max = 10000) {
+const getRandomInt = (min = 11, max = 10000) => {
     min = Math.ceil(min);
     max = Math.floor(max);
-    return Math.floor(Math.random() * (max - min)) + min; //Максимум не включается, минимум включается
-}
+    return Math.floor(Math.random() * (max - min)) + min;
+};
 
 module.exports = getRandomInt;
